+++ src/os2datascanner/projects/admin/adminapp/models/scannerjobs/googledrivescanner.py
@@ -7,6 +7,7 @@ from django.conf import settings
 from django.core.exceptions import ValidationError
 from django.utils.translation import pgettext_lazy
 
+from django.utils.translation import gettext_lazy as _
 from .scanner import Scanner
 from ...utils import upload_path_gdrive_users, upload_path_gdrive_service_account
 from os2datascanner.engine2.model.googledrive import GoogleDriveSource
@@ -57,3 +58,6 @@ class GoogleDriveScanner(Scanner):
 
     object_name = pgettext_lazy("unit of scan", "file")
     object_name_plural = pgettext_lazy("unit of scan", "files")
+
+    class Meta:
+        verbose_name = _("Google Drive scanner")