+++ src/os2datascanner/projects/admin/adminapp/models/scannerjobs/sbsysscanner.py
@@ -1,5 +1,6 @@
 from django.conf import settings
 from django.utils.translation import pgettext_lazy
+from django.utils.translation import gettext_lazy as _
 from .scanner import Scanner
 
 from os2datascanner.engine2.model.sbsys import SbsysSource
@@ -21,3 +22,6 @@ class SbsysScanner(Scanner):
 
     object_name = pgettext_lazy("unit of scan", "case")
     object_name_plural = pgettext_lazy("unit of scan", "cases")
+
+    class Meta:
+        verbose_name = _("SBSYS scanner")