+++ src/os2datascanner/projects/admin/adminapp/urls.py
@@ -41,7 +41,8 @@ from .views.rule_views import (RuleList, CustomRuleCreate,
 
 from .views.scanner_views import (StatusOverview, StatusCompletedView, StatusCompletedCSVView,
                                   StatusDelete, StatusTimeline, UserErrorLogView,
-                                  UserErrorLogCSVView)
+                                  UserErrorLogCSVView, RemovedScannersView, RecreateScannerView,
+                                  DeleteRemovedScannerView)
 
 
 from .views.miniscanner_views import MiniScanner, execute_mini_scan
@@ -72,6 +73,11 @@ urlpatterns = [
     re_path(r'^status/(?P<pk>\d+)/delete/$', StatusDelete.as_view(), name='status-delete'),
     re_path(r'^help/guide/$', GuideView.as_view(), name='guide'),
     re_path(r'^org-units-listing/', OrganizationalUnitListing.as_view(), name='org-units-listing'),
+    path("scanners/removed/", RemovedScannersView.as_view(), name="removed_scanners"),
+    path("scanners/removed/<int:pk>/recreate", RecreateScannerView.as_view(),
+         name="recreate_scanner"),
+    path("scanners/removed/<int:pk>/delete", DeleteRemovedScannerView.as_view(),
+         name="delete_removed_scanner"),
 
 
     re_path(r'^(msgraphcalendarscanners|msgraphmailscanners|msgraphfilescanners|'